+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## [0.3.0] - 2022-03-25
+### Changes
+- Add MMD shader to UniToon shader converter.
+- Fix lighting logic.
+
 ## [0.2.0] - 2022-03-24
 ### Changes
 - Separate files by version.